+++ test/Test/QuickCheck/Extra.hs
@@ -3,7 +3,9 @@ module Test.QuickCheck.Extra
   , (>=!)
   ) where
 
-import Test.QuickCheck (Property, counterexample)
+import           Test.QuickCheck                ( Property
+                                                , counterexample
+                                                )
 
 infix 4 <=!
 (<=!) :: (Ord a, Show a) => a -> a -> Property
@@ -12,4 +14,3 @@ x <=! y = counterexample (show x ++ " ≰ " ++ show y) (x <= y)
 infix 4 >=!
 (>=!) :: (Ord a, Show a) => a -> a -> Property
 x >=! y = counterexample (show x ++ " ≱ " ++ show y) (x >= y)
-