+++ changelog.md
@@ -1,6 +1,8 @@
 # Change Log
 
 ## WIP
+  - Hyperbolic anomaly support
+  - Several new utility functions
 
 ## 0.3
   - Switch to `units` from `uom-plugin`